# openapi.yaml
openapi: 3.0.3
info:
  title: Coding Agent Platform APIs
  version: 0.1.0
  description: |
    Consolidated API specification for the Coding Agent platform.
    This document aggregates three backend services:
      - Model service (LLM inference)
      - Execution gateway (user-facing run API)
      - Runner service (isolated container execution)

    Note: When any endpoint is added/changed, update this file accordingly.

tags:
  - name: model
    description: LLM inference endpoints
  - name: execution
    description: User-facing code execution gateway
  - name: runner
    description: Low-level code runner (ephemeral containers)
  - name: meta
    description: Health, readiness, and metrics

paths:
  /:
    get:
      tags: [model, execution]
      summary: Root message
      description: Simple ping endpoint for model or execution service
      operationId: rootMessage
      servers:
        - url: http://localhost:8000
          description: model service
        - url: http://localhost:5000
          description: execution gateway
      responses:
        '200':
          description: Service greeting
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

  /healthz:
    get:
      tags: [meta]
      summary: Liveness probe
      operationId: healthz
      servers:
        - url: http://localhost:8000
          description: model service
        - url: http://localhost:5000
          description: execution gateway
        - url: http://localhost:5100
          description: runner service
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
        '503':
          description: Service unavailable (e.g., runner when Docker is down)

  /readyz:
    get:
      tags: [meta]
      summary: Readiness probe
      operationId: readyz
      servers:
        - url: http://localhost:8000
          description: model service
        - url: http://localhost:5000
          description: execution gateway
        - url: http://localhost:5100
          description: runner service
      responses:
        '200':
          description: Ready status
          content:
            application/json:
              schema:
                type: object
                properties:
                  ready:
                    type: boolean

  /metrics:
    get:
      tags: [meta]
      summary: Prometheus metrics
      operationId: metrics
      servers:
        - url: http://localhost:8000
          description: model service
        - url: http://localhost:5000
          description: execution gateway
        - url: http://localhost:5100
          description: runner service
      responses:
        '200':
          description: Prometheus text format
          content:
            text/plain:
              schema:
                type: string

  /generate_code/:
    post:
      tags: [model]
      summary: Generate code from a prompt
      operationId: generateCode
      servers:
        - url: http://localhost:8000
          description: model service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PromptRequest'
      responses:
        '200':
          description: Generated code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateCodeResponse'
        '200_error':
          description: Model not ready (compat)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /generate_code_stream:
    post:
      tags: [model]
      summary: Generate code (streaming)
      description: Streams generated text tokens as they are produced
      operationId: generateCodeStream
      servers:
        - url: http://localhost:8000
          description: model service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PromptRequest'
      responses:
        '200':
          description: Chunked text stream
          content:
            text/plain:
              schema:
                type: string
              examples:
                chunked:
                  value: "partial token stream..."

  /execute_code/:
    post:
      tags: [execution]
      summary: Execute user code (gateway)
      operationId: executeCode
      servers:
        - url: http://localhost:5000
          description: execution gateway
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CodeRequest'
      responses:
        '200':
          description: Execution result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecuteResponse'

  /execute_code_stream/:
    post:
      tags: [execution]
      summary: Execute user code (streaming)
      operationId: executeCodeStream
      servers:
        - url: http://localhost:5000
          description: execution gateway
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CodeRequest'
      responses:
        '200':
          description: Chunked logs stream; ends with a JSON line containing exit_code and duration_ms
          content:
            text/plain:
              schema:
                type: string

  /run:
    post:
      tags: [runner]
      summary: Execute code in an ephemeral container
      operationId: runnerRun
      servers:
        - url: http://localhost:5100
          description: runner service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunRequest'
      responses:
        '200':
          description: Runner result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunnerResponse'
        '500':
          description: Runner or Docker error

  /run_stream:
    post:
      tags: [runner]
      summary: Execute code in an ephemeral container (streaming)
      operationId: runnerRunStream
      servers:
        - url: http://localhost:5100
          description: runner service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunRequest'
      responses:
        '200':
          description: Chunked log stream ending with a JSON line containing exit_code and duration_ms
          content:
            text/plain:
              schema:
                type: string

components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string

    PromptRequest:
      type: object
      properties:
        prompt:
          type: string
        max_new_tokens:
          type: integer
          minimum: 1
          maximum: 1024
          default: 256
        temperature:
          type: number
          minimum: 0
          default: 0.2
        top_p:
          type: number
          minimum: 0
          maximum: 1
          default: 0.95
        stop:
          type: array
          items:
            type: string

    GenerateCodeResponse:
      type: object
      properties:
        generated_code:
          type: string
      required: [generated_code]

    CodeRequest:
      type: object
      properties:
        code:
          type: string
        language:
          type: string
          enum: [python]
          default: python
        timeout_seconds:
          type: integer
          minimum: 1
          maximum: 120
          default: 10
        memory_mb:
          type: integer
          minimum: 64
          maximum: 2048
          default: 256
        cpu_millis:
          type: integer
          minimum: 100
          maximum: 2000
          default: 500
      required: [code]

    ExecuteResponse:
      type: object
      properties:
        logs:
          type: string
        exit_code:
          type: integer
        duration_ms:
          type: integer

    RunRequest:
      type: object
      properties:
        code:
          type: string
        language:
          type: string
          enum: [python]
          default: python
        timeout_seconds:
          type: integer
          minimum: 1
          maximum: 120
          default: 10
        memory_mb:
          type: integer
          minimum: 64
          maximum: 2048
          default: 256
        cpu_millis:
          type: integer
          minimum: 100
          maximum: 2000
          default: 500
      required: [code]

    RunnerResponse:
      type: object
      properties:
        exit_code:
          type: integer
        logs:
          type: string
        duration_ms:
          type: integer